+++ debezium-scripting/debezium-scripting/src/test/resources/wasm/go/filter5.go
@@ -6,8 +6,8 @@ import (
 
 //export process
 func process(proxyPtr uint32) uint32 {
-	var valueId = debezium.GetInt(debezium.Get(proxyPtr, "value.after.id"))
-	var valueLsn = debezium.GetInt(debezium.Get(proxyPtr, "value.source.lsn"))
+	var valueId = debezium.GetInt8(debezium.Get(proxyPtr, "value.after.id"))
+	var valueLsn = debezium.GetInt32(debezium.Get(proxyPtr, "value.source.lsn"))
 	var valueVersion = debezium.GetString(debezium.Get(proxyPtr, "value.source.version"))
 	var topicName = debezium.GetString(debezium.Get(proxyPtr, "topic"))
